Refactor sync-full command to utilize worker threads for message synchronization
- Implemented worker threads to parallelize message fetching and processing, improving performance and efficiency.
- Enhanced progress tracking during synchronization, providing real-time updates on completion status.
- Introduced a new sync-worker module to handle individual channel processing, ensuring better resource management.
- Updated database write logic to handle batch inserts for aggregated message counts, optimizing database interactions.
- Made minor adjustments to the write-behind mechanism for improved data consistency.

+++ src/database/write-behind.ts
@@ -92,9 +92,9 @@ export async function flushDirtyCountsToDB() {
         const query = `
           INSERT INTO message_counts (user_id, guild_id, channel_id, message_date, count)
           SELECT
-            t.user_id,
-            $1 AS guild_id,
-            t.channel_id,
+            t.user_id::bigint,
+            $1::bigint AS guild_id,
+            t.channel_id::bigint,
             t.message_date::date,
             t.count
           FROM UNNEST(
